Refactor service specs
