+++ docker/services/app/workflows/handle_switch_event.rb
@@ -7,65 +7,15 @@ class HandleSwitchEvent < ApplicationWorkflow
 
   def call
     if event.task_running? && event.eni_attached?
-      create_opensips_load_balancer_targets!
+      load_balancer_manager.create_targets
     elsif event.task_stopped? && event.eni_deleted?
-      delete_opensips_load_balancer_targets!
+      load_balancer_manager.delete_targets
     end
   end
 
   private
 
-  def create_opensips_load_balancer_targets!
-    gateway_databases.each do |database_connection|
-      database_connection.transaction do
-        load_balancer_targets.each do |load_balancer_target|
-          create_opensips_load_balancer_target!(load_balancer_target:, database_connection:)
-        end
-      end
-    end
-  end
-
-  def delete_opensips_load_balancer_targets!
-    gateway_databases.each do |database_connection|
-      OpenSIPSLoadBalancerTarget.where(
-        dst_uri: load_balancer_targets.map(&:dst_uri),
-        database_connection:
-      ).delete
-    end
-  end
-
-  def create_opensips_load_balancer_target!(load_balancer_target:, database_connection:)
-    return if OpenSIPSLoadBalancerTarget.exists?(dst_uri: load_balancer_target.dst_uri, database_connection:)
-
-    OpenSIPSLoadBalancerTarget.new(
-      dst_uri: load_balancer_target.dst_uri,
-      resources: load_balancer_target.resources,
-      group_id: 1,
-      probe_mode: 2,
-      database_connection:
-    ).save!
-  end
-
-  def gateway_databases
-    @gateway_databases ||= DatabaseConnections.gateways
-  end
-
-  def load_balancer_targets
-    @load_balancer_targets ||= [
-      build_load_balancer_target(port: fs_sip_port, resources_identifier: "gw"),
-      build_load_balancer_target(port: fs_sip_alternative_port, resources_identifier: "gwalt")
-    ]
-  end
-
-  def build_load_balancer_target(port:, resources_identifier:)
-    LoadBalancerTarget.new(ip_address: event.eni_private_ip, port:, resources_identifier:)
-  end
-
-  def fs_sip_port
-    ENV.fetch("FS_SIP_PORT", "5060")
-  end
-
-  def fs_sip_alternative_port
-    ENV.fetch("FS_SIP_ALTERNATIVE_PORT", "5080")
+  def load_balancer_manager
+    @load_balancer_manager ||= ManageLoadBalancerTargets.new(ip_address: event.private_ip)
   end
 end
